Various things

+++ include/algorithm/node.h
@@ -56,7 +56,7 @@ void node_init(Node *node, int n);
 
 char nodes_eq(Node *n1, Node *n2);
 
-LocalNode node_local_alloc(int id, int n, double hb_timeout);
+LocalNode node_local_alloc(int id, int n, long hb_timeout);
 
 void node_local_dealloc(LocalNode *n);
